Fifth commit
added option to Modify password

added the whole password strength system for indicating password strength and categorize it by strength

added text field for easy password and username copy options.

started with the Database section using MySQ, for the moment it only saves on exit and read on initialize.

changed font for many parts of the software

added instructions about the software and about the password strength category system.